+++ .github/release-notes.md
@@ -1,2 +1,4 @@
 * Fixed toasts not being interactable with a modal open
 * Fixed toasts duplicating when changing pages
+* Added context.watchConsole function for scripter
+* Automatically create scripts directory